requirements.txt
feedparser==6.0.11
httpx==0.27.0
openai==1.30.0
pyyaml==6.0.1
python-dotenv==1.0.1
anthropic==0.28.0
